refactor and split reasoning test
